requirements.txt
altair==5.1.2
ansi2html==1.8.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autoviz==0.1.732
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
bokeh==2.4.3
branca==0.7.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
colorcet==3.0.1
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
dash==2.14.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dataviz==1.1.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
emoji==2.8.0
executing==2.0.1
fastjsonschema==2.19.0
fiona==1.9.5
Flask==3.0.0
folium==0.15.0
fonttools==4.44.3
fqdn==1.5.1
fsspec==2023.10.0
geopandas==0.14.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
holoviews==1.14.9
hvplot==0.7.3
idna==3.4
importlib-metadata==6.8.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.11.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.10.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.1
kiwisolver==1.4.5
lab==8.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
overrides==7.4.0
packaging==23.2
pandas==2.1.3
pandas-dq==1.28
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
patsy==0.5.3
Pillow==10.1.0
platformdirs==4.0.0
plotly==5.18.0
plotly-express==0.4.1
prometheus-client==0.18.0
prompt-toolkit==3.0.41
protobuf==4.25.1
psutil==5.9.6
pyamg==5.0.1
pyarrow==14.0.1
pycparser==2.21
pyct==0.5.0
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pyviz_comms==3.0.0
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.0
regex==2023.10.3
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.12.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
Send2Trash==1.8.2
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.28.2
streamlit-dynamic-filters==0.1.3
tenacity==8.2.3
terminado==0.18.0
textblob==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
txt2tags==3.9
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.10
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wordcloud==1.9.2
xgboost==2.0.2
xlrd==2.0.1
xyzservices==2023.10.1
zipp==3.17.0